+++ spec/lib/karafka/swarm/liveness_listener_spec.rb
@@ -14,27 +14,27 @@ RSpec.describe_current do
 
   after { Karafka::App.config.swarm.node = false }
 
-  context "when statistics are emitted too often" do
+  context "when events poll happens too often" do
     context "when node becomes orphaned" do
       let(:orphaned) { [false, true] }
 
       before do
-        listener.on_statistics_emitted(nil)
+        listener.on_client_events_poll(nil)
         allow(Kernel).to receive(:exit!)
       end
 
       it "expect to do nothing as too often" do
-        listener.on_statistics_emitted(nil)
+        listener.on_client_events_poll(nil)
         expect(Kernel).not_to have_received(:exit!)
       end
     end
   end
 
-  context "when statistics are emitted not too often" do
+  context "when events poll happens not too often" do
     before { allow(node).to receive(:write) }
 
     it "expect to write ok status to the node" do
-      listener.on_statistics_emitted(nil)
+      listener.on_client_events_poll(nil)
       expect(node).to have_received(:write).with("0")
     end
 
@@ -44,7 +44,7 @@ RSpec.describe_current do
       before { allow(Kernel).to receive(:exit!) }
 
       it "expect to exit!" do
-        listener.on_statistics_emitted(nil)
+        listener.on_client_events_poll(nil)
         expect(Kernel).to have_received(:exit!).with(3)
       end
     end